Added feature to display current team members Moved functionality up from the add-project-agents component to add-project so the state of the project can be updated
